+++ lib/voicemail/call_controllers/mailbox_messages_controller.rb
@@ -27,7 +27,7 @@ module Voicemail
     end
 
     def handle_message(message)
-      invoke MailboxPlayMessageController, message: message, mailbox: mailbox[:id]
+      invoke MailboxPlayMessageController, message: message, mailbox: mailbox[:id], new_or_saved: new_or_saved
     end
 
     private