+++ lib/voicemail/plugin.rb
@@ -36,7 +36,7 @@ module Voicemail
         number_before "You have ", desc: "What to play before the number of new messages"
         number_after_new " new messages", desc: "What to play after the number of new messages"
         number_after_saved " saved messages", desc: "What to play after the number of saved messages"
-        menu_greeting "Press 1 to listen to new messages, 2 to change your greeting, 3 to change your PIN", desc: "What to say before the main menu"
+        menu_greeting "Press 1 to listen to new messages, 2 to listen to saved messages, 3 to change your greeting, 4 to change your PIN", desc: "What to say before the main menu"
         menu_timeout_message "Please enter a digit for the menu", desc: "Message to play on main menu timeout"
         menu_invalid_message "Please enter valid input", desc: "Message to play on main menu invalid"
         menu_failure_message "Sorry, unable to understand your input.", desc: "Message to play on main menu failure"
@@ -67,7 +67,8 @@ module Voicemail
 
       desc "Listen to messages menu configuration"
       messages {
-        menu "Press 1 to archive the message and go to the next, press 5 to delete the message and go to the next, press 7 to hear the message again, press 9 for the main menu", desc: "Menu to use inside messages"
+        menu_new "Press 1 to archive the message and go to the next, press 5 to delete the message and go to the next, press 7 to hear the message again, press 9 for the main menu", desc: "Menu to use inside new messages"
+        menu_saved "Press 1 to unarchive the message and go to the next, press 5 to delete the message and go to the next, press 7 to hear the message again, press 9 for the main menu", desc: "Menu to use inside saved messages"
         no_new_messages "There are no new messages", desc: "Message to inform the user he has no new messages"
         no_saved_messages "There are no saved messages", desc: "Message to inform the user he has no saved messages"
         message_received_on "Message received on ", desc: "Prefix to menu intro"