+++ spec/voicemail/call_controllers/mailbox_messages_controller_spec.rb
@@ -66,7 +66,7 @@ module Voicemail
 
       describe "#handle_message" do
         it "invokes MailboxPlayMessageController" do
-          should_invoke Voicemail::MailboxPlayMessageController, message: message, mailbox: mailbox[:id]
+          should_invoke Voicemail::MailboxPlayMessageController, message: message, mailbox: mailbox[:id], new_or_saved: :new
           controller.handle_message message
         end
       end
@@ -101,7 +101,7 @@ module Voicemail
 
       describe "#handle_message" do
         it "invokes MailboxPlayMessageController" do
-          should_invoke Voicemail::MailboxPlayMessageController, message: message, mailbox: mailbox[:id]
+          should_invoke Voicemail::MailboxPlayMessageController, message: message, mailbox: mailbox[:id], new_or_saved: :saved
           controller.handle_message message
         end
       end